+++ custom_components/eshtaya_smart_control/panel.py
@@ -25,10 +25,10 @@ async def async_register_panel(hass: HomeAssistant) -> None:
         hass=hass,
         frontend_url_path=PANEL_URL,
         webcomponent_name=PANEL_ELEMENT,
-        module_url=f"{STATIC_URL}/smart-control-panel.js?v={VERSION}",
+        module_url=f"{STATIC_URL}/smart-control-panel-v21.js?v={VERSION}",
         sidebar_title=PANEL_TITLE,
         sidebar_icon=PANEL_ICON,
-        require_admin=True,
+        require_admin=False,
         config_panel_domain=DOMAIN,
         config={},
     )